move watch logic (mostly) to cli-lib
