+++ packages/cli-lib/lang/en.lyaml
@@ -491,19 +491,20 @@ en:
             examples:
               default: "Watch a project within the myProjectFolder folder"
             logs: 
-              watching: "Watcher is ready and watching {{ projectDir }}. Any changes detected will be automatically uploaded and added to the current staging build."
+              watching: "Watcher is ready and watching \"{{ projectDir }}\". Any changes detected will be automatically uploaded and added to the current staging build."
               resuming: "Resuming watcher..."
-              uploadSucceeded: "Uploaded file {{ filePath }} to {{ remotePath }}"
+              uploadSucceeded: "Uploaded file \"{{ filePath }}\" to \"{{ remotePath }}\""
               createNewBuild: "Staging a new build..."
               buildCancelled: "Staging build has been cancelled. Please try running `hs project watch` again"
             debug:
-              uploadFailed: "Failed to upload file {{ filePath }} to {{ remotePath }}"
+              uploadFailed: "Failed to upload file \"{{ filePath }}\" to \"{{ remotePath }}\""
               uploading: "Attempting to upload file \"{{ filePath }}\" to \"{{ remotePath }}\""
-              extensionNotAllowed: "Skipping {{ filePath }} due to unsupported extension"
-              ignored: "Skipping {{ filePath }} due to an ignore rule"
+              extensionNotAllowed: "Skipping \"{{ filePath }}\" due to unsupported extension"
+              ignored: "Skipping \"{{ filePath }}\" due to an ignore rule"
               pause: "Pausing watcher, attempting to queue build"
               buildStarted: "Build queued."
               buildCancelled: "The current staging build has been cancelled. Exiting..."
+              fileAlreadyQueued: "File \"{{ filePath }}\" is already queued for upload"
             errors:
               projectLocked: "The previous staging build is still in progress"
       remove:
